// src/model/typescript/swissrets-model.ts
export interface SwissRetsInventory {
  created?: string;
  generator: Generator;
  projects?: Project[];
  properties?: Property[];
}

export interface Generator {
  name?: string;
  version: string;
}

/**
 * Building project container.
 */
export interface Project {
  /**
   * Location of the property.
   */
  address?: Address;
  /**
   * Mutually exclusive lifecycle state.
   */
  availability: ProjectAvailability;
  /**
   * Characteristics and ranges relevant to entire project.
   */
  characteristics?: ProjectCharacteristics;
  /**
   * Construction status of project.
   */
  constructionStatus?: ConstructionStatus;
  id?: string;
  /**
   * Sequence of all language specific content and texts.
   */
  localizations?: ProjectLocalization[];
  /**
   * Price ranges for marketing purposes.
   */
  prices?: ProjectPrices;
  referenceId: string;
  /**
   * The owner or the owners substitute (for example a broker).
   */
  seller?: ProjectSeller;
  /**
   * List of units.
   */
  units: Unit[];
}

/**
 * Location of the property.
 *
 * Postal address.
 */
export interface Address {
  /**
   * Uppercase two letter country code.
   */
  countryCode?: string;
  /**
   * Geo coordinates and elevation
   */
  geo?: Geo;
  locality?: string;
  postalCode?: string;
  postOfficeBoxNumber?: string;
  region?: string;
  street?: string;
  streetAddition?: string;
  streetNumber?: string;
  subunit?: number;
}

/**
 * Geo coordinates and elevation
 */
export interface Geo {
  elevation?: number;
  latitude: number;
  longitude: number;
}

/**
 * Mutually exclusive lifecycle state.
 */
export interface ProjectAvailability {
  expiration?: string;
  start?: string;
  state: State;
}

/**
 * Lifecycle state of properties.
 */
export enum State {
  Active = 'active',
  Private = 'private',
  Reference = 'reference',
  Reserved = 'reserved',
  Taken = 'taken'
}

/**
 * Characteristics and ranges relevant to entire project.
 */
export interface ProjectCharacteristics {
  areaBwfFrom?: number;
  areaBwfTo?: number;
  areaSiaNfFrom?: number;
  areaSiaNfTo?: number;
  numberOfCommercialUnits?: number;
  numberOfResidentialUnits?: number;
  numberOfRoomsFrom?: number;
  numberOfRoomsTo?: number;
}

/**
 * Construction status of project.
 */
export enum ConstructionStatus {
  Completed = 'completed',
  Planned = 'planned',
  Ready = 'ready',
  UnderConstruction = 'under-construction'
}

/**
 * One for each language, a set of language specific content and texts.
 */
export interface ProjectLocalization {
  /**
   * Sequence links, files, embeds and media.
   */
  attachments?: ProjectAttachment;
  /**
   * Main description.
   */
  description?: string;
  /**
   * A short extract from description, location and equipment.
   */
  excerpt?: string;
  languageCode: string;
  /**
   * Simple marketing title.
   */
  title: string;
}
// ...
export interface ProjectPrices {
  /**
   * One time buy prices.
   */
  buy?: ProjectPriceBuy;
  currency: string;
  /**
   * Recurring rental price.
   */
  rent?: ProjectPriceRent;
}
// ...
export interface ProjectSeller {
  /**
   * Contact person for questions and typically responsible for making appointments. This
   * contact is typicaly presented publicly.
   */
  contactPerson?: Person;
  /**
   * The person responsible for receiving inquiries and applications. This person should not
   * be presented publicly and is only intended as an indirect recipient using an online form
   * or similar.
   */
  inquiryPerson?: Person;
  /**
   * The company selling
   */
  organization?: Organization;
}
// ...
export interface Unit {
  characteristics?: UnitCharacteristic[];
  id?: string;
  /**
   * Sequence of all language specific content and texts.
   */
  localizations: UnitLocalization[];
  referenceId: string;
}

/**
 * The cold facts, to represent a unit.
 */
export interface UnitCharacteristic {
  numberOfFloors?: number;
}

/**
 * One for each language, a set of language specific content and texts.
 */
export interface UnitLocalization {
  languageCode: string;
  /**
   * Simple marketing title for unit.
   */
  title: string;
}

export interface Property {
  /**
   * Location of the property.
   */
  address?: Address;
  /**
   * The name or username of the author, who created the record.
   */
  author?: string;
  /**
   * Mutually exclusive lifecycle state.
   */
  availability: PropertyAvailability;
  /**
   * Federal statistical office specific information
   */
  bfs?: Bfs;
  /**
   * Building zones intended for residential buildings or for other buildings intended as
   * long-stay accommodation, those are typically municipal-specific abbreviations.
   */
  buildingZones?: string;
  /**
   * Sequence of ordered categories (main first).
   */
  categories?: Category[];
  /**
   * Main characteristics of properties
   */
  characteristics?: PropertyCharacteristics;
  /**
   * Creation date and time of the data.
   */
  created?: string;
  /**
   * The development state of the property.
   */
  development?: Development;
  /**
   * Object references based on other than SwissRETS standard (e.g IDX).
   */
  externalReference?: ExternalReference;
  /**
   * The main heating system.
   */
  heating?: Heating;
  /**
   * This ID needs to be unique and represents the id from the exporter (this can be the same
   * as referenceId if the exporter is the original creator)
   */
  id: string;
  /**
   * Sequence of all language specific content and texts.
   */
  localizations: PropertyLocalization[];
  /**
   * Quality label for new and refurbished low-energy-consumption buildings.
   */
  minergieCertification?: MinergieCertification;
  /**
   * Date and time of the last modification. This field should only be used for presentational
   * purposes and should not be relied upon for import cache-busting.
   */
  modified?: string;
  /**
   * Arbitrary string that holds one or more parcel numbers, either from the swiss cadastral
   * system or from a certain municipality.
   */
  parcelNumbers?: string;
  prices?: PropertyPrices;
  /**
   * List of publishers
   */
  publishers?: Publisher[];
  /**
   * Can be used to track the original creation id (this should usually be unique to the
   * property)
   */
  referenceId: string;
  seller?: PropertySeller;
  /**
   * Specifies rent or buy.
   */
  type: PropertyType;
  /**
   * Reference to an existing project - unit within the xml. Used for grouping and
   * representing a 'project-tree'
   */
  unitReferenceId?: string;
  /**
   * Prioritized sequence for the intended utilization of the property, main comes first.
   */
  utilizations?: Utilization[];
  /**
   * Can should only be used for presentational purposes publishers tend to present the
   * referenceId if this one is missing
   */
  visualReferenceId?: string;
}

/**
 * Mutually exclusive lifecycle state.
 */
export interface PropertyAvailability {
  expiration?: string;
  start?: string;
  state: State;
}

/**
 * Federal statistical office specific information
 */
export interface Bfs {
  /**
   * The building identifier of the swiss population register
   */
  egid?: string;
  /**
   * The property (plot) identifier of the swiss population register
   */
  egrid?: string;
  /**
   * The flat identifier of the swiss population register
   */
  ewid?: string;
}

export enum Category {
  AdvertisingArea = 'advertising-area',
  AgriculturalLot = 'agricultural-lot',
  AllotmentGarden = 'allotment-garden',
  Apartment = 'apartment',
  Arcade = 'arcade',
  Atelier = 'atelier',
  AtticCompartment = 'attic-compartment',
  AtticFlat = 'attic-flat',
  Bakery = 'bakery',
  Bar = 'bar',
  BoatDryDock = 'boat-dry-dock',
  BoatLandingStage = 'boat-landing-stage',
  BoatMooring = 'boat-mooring',
  BuildingLot = 'building-lot',
  Butcher = 'butcher',
  Cafe = 'cafe',
  Campground = 'campground',
  CarPark = 'car-park',
  CarRepairShop = 'car-repair-shop',
  CarpentryShop = 'carpentry-shop',
  Casino = 'casino',
  Castle = 'castle',
  CellarCompartment = 'cellar-compartment',
  Chalet = 'chalet',
  CheeseFactory = 'cheese-factory',
  ClubDisco = 'club-disco',
  CommercialLot = 'commercial-lot',
  CommercialSpace = 'commercial-space',
  CoveredMotorbikeParkingSpace = 'covered-motorbike-parking-space',
  CoveredParkingSpace = 'covered-parking-space',
  DepartmentStore = 'department-store',
  DetachedHouse = 'detached-house',
  DisplayWindow = 'display-window',
  DoctorsOffice = 'doctors-office',
  DoubleGarage = 'double-garage',
  DuplexHouse = 'duplex-house',
  EarthShelteredDwelling = 'earth-sheltered-dwelling',
  Factory = 'factory',
  FarmHouse = 'farm-house',
  FuelStation = 'fuel-station',
  FurnishedFlat = 'furnished-flat',
  GolfCourse = 'golf-course',
  GrannyFlat = 'granny-flat',
  Hairdresser = 'hairdresser',
  HobbyRoom = 'hobby-room',
  HorseBox = 'horse-box',
  Hospital = 'hospital',
  Hotel = 'hotel',
  IndoorSwimmingPool = 'indoor-swimming-pool',
  IndoorTennisCourt = 'indoor-tennis-court',
  IndustrialLot = 'industrial-lot',
  IndustrialObject = 'industrial-object',
  Kiosk = 'kiosk',
  Laboratory = 'laboratory',
  Library = 'library',
  Loft = 'loft',
  Maisonette = 'maisonette',
  MarketGarden = 'market-garden',
  MiniGolfCourse = 'mini-golf-course',
  Motel = 'motel',
  MovieTheater = 'movie-theater',
  MultiplexHouse = 'multiplex-house',
  MultistoreyCarPark = 'multistorey-car-park',
  NursingHome = 'nursing-home',
  Office = 'office',
  OneRoomFlat = 'one-room-flat',
  OpenMotorbikeParkingSpace = 'open-motorbike-parking-space',
  OpenParkingSpace = 'open-parking-space',
  Orphanage = 'orphanage',
  OutdoorSwimmingPool = 'outdoor-swimming-pool',
  ParkingSpace = 'parking-space',
  PartyRoom = 'party-room',
  Plot = 'plot',
  Pub = 'pub',
  ResidentialCommercialBuilding = 'residential-commercial-building',
  Restaurant = 'restaurant',
  RetailLocation = 'retail-location',
  RetirementHome = 'retirement-home',
  RidingHall = 'riding-hall',
  RoofFlat = 'roof-flat',
  RowHouse = 'row-house',
  Rustico = 'rustico',
  Sanatorium = 'sanatorium',
  Sauna = 'sauna',
  ShoppingCenter = 'shopping-center',
  SingleGarage = 'single-garage',
  SingleRoom = 'single-room',
  Solarium = 'solarium',
  SportsHall = 'sports-hall',
  SquashBadminton = 'squash-badminton',
  SteppedFlat = 'stepped-flat',
  SteppedHouse = 'stepped-house',
  Stoeckli = 'stoeckli',
  StorageRoom = 'storage-room',
  TennisCourt = 'tennis-court',
  UndergroundParkingSpace = 'underground-parking-space',
  Villa = 'villa',
  Workshop = 'workshop'
}
// ...
export enum Development {
  Full = 'full',
  Partial = 'partial',
  Undeveloped = 'undeveloped'
}

/**
 * Object references based on other than SwissRETS standard (e.g IDX).
 */
export interface ExternalReference {
  refHouse?: string;
  refObject?: string;
  refProperty?: string;
}

/**
 * The main heating system.
 */
export interface Heating {
  /**
   * How the heating system distributes and maintains temperature.
   */
  distribution?: HeatingDistribution;
  /**
   * Technology used to create heat.
   */
  generation?: HeatingGeneration;
}
// ...
export interface PropertyLocalization {
  /**
   * Sequence links, files, embeds and media.
   */
  attachments?: PropertyAttachment;
  /**
   * Main description.
   */
  description?: string;
  /**
   * Description of available and included equipment.
   */
  equipment?: string;
  /**
   * List of events
   */
  events?: Event[];
  /**
   * A short extract from description, location and equipment.
   */
  excerpt?: string;
  languageCode: string;
  /**
   * Description of the surrounding and location.
   */
  location?: string;
  /**
   * Simple marketing title.
   */
  title: string;
  /**
   * Specific visiting instructions or information, where to get the key, when to contact and
   * similar.
   */
  visitInformation?: string;
}
// ...
export enum MinergieCertification {
  Minergie = 'Minergie',
  MinergieA = 'Minergie-A',
  MinergieAEco = 'Minergie-A-Eco',
  MinergieAreal = 'Minergie-Areal',
  MinergieEco = 'Minergie-Eco',
  MinergieMqsBau = 'Minergie-Mqs-Bau',
  MinergieMqsBetrieb = 'Minergie-Mqs-Betrieb',
  MinergieP = 'Minergie-P',
  MinergiePEco = 'Minergie-P-Eco'
}

/**
 * Sell, rent, deposit and auction prices.
 */
export interface PropertyPrices {
  /**
   * List of extras.
   */
  additionalOffers?: SchemaJson[];
  /**
   * Starting price for an auction.
   */
  auction?: Auction;
  /**
   * One time buy prices.
   */
  buy?: PropertyPriceBuy;
  currency?: string;
  /**
   * Required amount of deposit.
   */
  deposit?: Deposit;
  /**
   * Recurring rental price.
   */
  rent?: PropertyPriceRent;
}
// ...
export interface Publisher {
  id: string;
  /**
   * List of publisher specific custom fields
   */
  options?: PublisherOption[];
  /**
   * List of promotions
   */
  promotions?: PublisherPromotion[];
}
// ...
export interface PropertySeller {
  /**
   * Contact person for questions and typically responsible for making appointments. This
   * contact is typicaly presented publicly.
   */
  contactPerson?: Person;
  inquiryEmail?: string;
  /**
   * The company selling
   */
  organization?: Organization;
  /**
   * Person responsible opening the door for visiting.
   */
  visitPerson?: Person;
}

/**
 * Specifies rent or buy.
 */
export enum PropertyType {
  Buy = 'buy',
  Rent = 'rent'
}